+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+* Fixed: Connection string not in result object
+
 ## 0.2.0 - 2020-05-17
 
 * Changed: Add quiet option for Test-SqlConnection command